requirements.txt
django==3.0.3
# DRF
requests==2.22.0
djangorestframework==3.12.1
django-cors-headers==3.5.0
drf-yasg==1.20.0
# Postgres
psycopg2==2.8.6
psycopg2-binary==2.8.6
# Celery
celery==5.0.5
django-redis==4.12.1
django-celery-beat==2.0.0
django-celery-results==2.0.1
# Docker
gunicorn==20.0.4


